Move purge outside ingest
